requirements.txt
ema-pytorch
tqdm
hydra-core
omegaconf
lpips
plyfile
timm
einops
imageio
wandb
moviepy
lightning==2.0
markupsafe==2.0.1
gradio
huggingface_hub